moved util to event utils
https://coveord.atlassian.net/browse/KIT-1525
